+++ Block/Menu.php
@@ -2,9 +2,11 @@
 
 namespace Snowdog\Menu\Block;
 
+use Magento\Customer\Model\Session;
 use Magento\Framework\Api\Search\FilterGroupBuilder;
 use Magento\Framework\Api\Search\SearchCriteriaFactory;
 use Magento\Framework\App\Cache\Type\Block;
+use Magento\Framework\App\Config\ScopeConfigInterface;
 use Magento\Framework\DataObject;
 use Magento\Framework\View\Element\Template;
 use Magento\Framework\Event\Manager as EventManager;
@@ -86,6 +88,16 @@ class Menu extends Template implements DataObject\IdentityInterface
     private $escaper;
 
     /**
+     * @var Session
+     */
+    private $customerSession;
+
+    /**
+     * @var ScopeConfigInterface
+     */
+    private $scopeConfig;
+
+    /**
      * @SuppressWarnings(PHPMD.ExcessiveParameterList)
      */
     public function __construct(
@@ -99,6 +111,8 @@ class Menu extends Template implements DataObject\IdentityInterface
         TemplateResolver $templateResolver,
         ImageFile $imageFile,
         Escaper $escaper,
+        Session $customerSession,
+        ScopeConfigInterface $scopeConfig,
         array $data = []
     ) {
         parent::__construct($context, $data);
@@ -113,6 +127,8 @@ class Menu extends Template implements DataObject\IdentityInterface
         $this->escaper = $escaper;
         $this->setTemplate($this->getMenuTemplate($this->_template));
         $this->submenuTemplate = $this->getSubmenuTemplate();
+        $this->customerSession = $customerSession;
+        $this->scopeConfig = $scopeConfig;
     }
 
     /**
@@ -410,7 +426,8 @@ class Menu extends Template implements DataObject\IdentityInterface
             ->setImageAltText($node->getImageAltText())
             ->setCustomTemplate($node->getNodeTemplate())
             ->setAdditionalData($node->getAdditionalData())
-            ->setSelectedItemId($node->getSelectedItemId());
+            ->setSelectedItemId($node->getSelectedItemId())
+            ->setCustomerGroups($node->getCustomerGroups());
 
         return $nodeBlock;
     }
@@ -442,10 +459,15 @@ class Menu extends Template implements DataObject\IdentityInterface
     private function fetchData()
     {
         $nodes = $this->nodeRepository->getByMenu($this->loadMenu()->getId());
+        $currentCustomerGroup = $this->getCustomerGroupId();
+        $customerGroupEnabled = $this->scopeConfig->getValue('snowmenu/general/customer_groups');
         $result = [];
         $types = [];
         foreach ($nodes as $node) {
             if (!$node->getIsActive()) {
+                continue;
+            }
+            if ($customerGroupEnabled && !$node->isVisible($currentCustomerGroup)) {
                 continue;
             }
 
@@ -501,5 +523,10 @@ class Menu extends Template implements DataObject\IdentityInterface
         }
 
         return $this->getMenuTemplate($baseSubmenuTemplate);
+    }
+
+    public function getCustomerGroupId()
+    {
+        return $this->customerSession->getCustomerGroupId();
     }
 }